api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx>=0.27.0
orjson==3.10.12
pyyaml==6.0.2
pytest==8.3.4
ruff==0.8.6

pytest-asyncio==0.24.0
